Low Priority cleanup: consolidated test helpers, removed redundant code
- Merged TestHelpers and SimpleHelpers into single test-helpers.ts
- Removed 1,365 lines of duplicate helper code
- Updated all test imports to use consolidated helpers
- Deleted auth-helpers.ts (79 lines, unused)
- Removed backwards compatibility fallback in middleware.ts
- Maintained all existing functionality while simplifying codebase
